Intercept all incoming request
Intercept all incoming request and extract JWT token from the header,
validate and set in execution context. It will enable JWT authentication
to all the end-points in out application and clients can access it using
the JWT token received from /authenticate end-point which itself
requires the default username and password to access it.